fix: sourcemaps for multiline strings
This deals with multiline strings simply by counting the number of new
lines in the string. Performance is a consideration in this library,
so I played around and determined this simple approach is, in addition
to being compact, the fastest versus a regex or split approach[1]

astring also did not test sourcemap generation with great thoroughness.
I've adjusted the generation test case to verify that at least one
mapping is generated for each test fixture, and that the line and
column is correct in the mappings.

Finally, I noticed that benchmarks were running in every test which
slowed my iteration down a bit. I added an environment variable
that can be set to skip these tests. I can remove it from this
PR if you don't want to take it.

Fixes #40

 1. https://gist.github.com/connor4312/0c534ae93356663a3cbad49c06d16d26